Scope CodeQL around intentional regex helper
